+++ saleor/graphql/menu/schema.py
@@ -26,16 +26,17 @@ class MenuQueries(graphene.ObjectType):
         MenuItem, query=graphene.String(description=DESCRIPTIONS['menu_item']),
         description='List of the shop\'s menu items.')
 
-    def resolve_menu(self, info, id=None, name=None):
-        return resolve_menu(info, id, name)
+    def resolve_menu(self, info, **data):
+        return resolve_menu(info, data.get('id'), data.get('name'))
 
-    def resolve_menus(self, info, query=None, **kwargs):
+    def resolve_menus(self, info, query=None, **_kwargs):
         return resolve_menus(info, query)
 
-    def resolve_menu_item(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, MenuItem)
+    def resolve_menu_item(self, info, **data):
+        return graphene.Node.get_node_from_global_id(
+            info, data.get('id'), MenuItem)
 
-    def resolve_menu_items(self, info, query=None, **kwargs):
+    def resolve_menu_items(self, info, query=None, **_kwargs):
         return resolve_menu_items(info, query)