+++ saleor/graphql/page/schema.py
@@ -3,7 +3,7 @@ import graphene
 from ..core.fields import PrefetchingConnectionField
 from ..descriptions import DESCRIPTIONS
 from ..translations.mutations import PageTranslate
-from .bulk_mutations import PageBulkDelete
+from .bulk_mutations import PageBulkDelete, PageBulkPublish, PageBulkUnpublish
 from .mutations import PageCreate, PageDelete, PageUpdate
 from .resolvers import resolve_page, resolve_pages
 from .types import Page
@@ -21,7 +21,7 @@ class PageQueries(graphene.ObjectType):
     def resolve_page(self, info, id=None, slug=None):
         return resolve_page(info, id, slug)
 
-    def resolve_pages(self, info, query=None, **kwargs):
+    def resolve_pages(self, info, query=None, **_kwargs):
         return resolve_pages(info, query=query)
 
 
@@ -29,5 +29,7 @@ class PageMutations(graphene.ObjectType):
     page_create = PageCreate.Field()
     page_delete = PageDelete.Field()
     page_bulk_delete = PageBulkDelete.Field()
+    page_bulk_publish = PageBulkPublish.Field()
+    page_bulk_unpublish = PageBulkUnpublish.Field()
     page_update = PageUpdate.Field()
     page_translate = PageTranslate.Field()