+++ saleor/graphql/product/types/products.py
@@ -29,7 +29,7 @@ from .attributes import Attribute, SelectedAttribute
 from .digital_contents import DigitalContent
 
 
-def prefetch_products(info, *args, **kwargs):
+def prefetch_products(info, *_args, **_kwargs):
     """Prefetch products visible to the current user.
 
     Can be used with models that have the `products` relationship. Queryset of
@@ -87,9 +87,10 @@ class ProductVariant(CountableDjangoObjectType):
         required=True, description='Quantity of a product available for sale.')
     price_override = graphene.Field(
         Money,
-        description=dedent("""Override the base price of a product if necessary.
-        A value of `null` indicates that the default product
-        price is used."""))
+        description=dedent(
+            """Override the base price of a product if necessary.
+               A value of `null` indicates that the default product
+               price is used."""))
     price = graphene.Field(Money, description='Price of the product variant.')
     attributes = graphene.List(
         graphene.NonNull(SelectedAttribute), required=True,
@@ -133,51 +134,51 @@ class ProductVariant(CountableDjangoObjectType):
         model = models.ProductVariant
 
     @permission_required('product.manage_products')
-    def resolve_digital_content(self, info):
+    def resolve_digital_content(self, *_args):
         return getattr(self, 'digital_content', None)
 
-    def resolve_stock_quantity(self, info):
+    def resolve_stock_quantity(self, *_args):
         return self.quantity_available
 
     @gql_optimizer.resolver_hints(
         prefetch_related='product__product_type__variant_attributes__values')
-    def resolve_attributes(self, info):
+    def resolve_attributes(self, *_args):
         attributes_qs = self.product.product_type.variant_attributes.all()
         return resolve_attribute_list(self.attributes, attributes_qs)
 
     @permission_required('product.manage_products')
-    def resolve_margin(self, info):
+    def resolve_margin(self, *_args):
         return get_margin_for_variant(self)
 
-    def resolve_price(self, info):
+    def resolve_price(self, *_args):
         return (
             self.price_override
             if self.price_override is not None else self.product.price)
 
     @permission_required('product.manage_products')
-    def resolve_price_override(self, info):
+    def resolve_price_override(self, *_args):
         return self.price_override
 
     @permission_required('product.manage_products')
-    def resolve_quantity(self, info):
+    def resolve_quantity(self, *_args):
         return self.quantity
 
     @permission_required(['order.manage_orders', 'product.manage_products'])
-    def resolve_quantity_ordered(self, info):
+    def resolve_quantity_ordered(self, *_args):
         # This field is added through annotation when using the
         # `resolve_report_product_sales` resolver.
         return getattr(self, 'quantity_ordered', None)
 
     @permission_required(['order.manage_orders', 'product.manage_products'])
-    def resolve_quantity_allocated(self, info):
+    def resolve_quantity_allocated(self, *_args):
         return self.quantity_allocated
 
     @permission_required(['order.manage_orders', 'product.manage_products'])
-    def resolve_revenue(self, info, period):
+    def resolve_revenue(self, *_args, period):
         start_date = reporting_period_to_date(period)
         return calculate_revenue_for_variant(self, start_date)
 
-    def resolve_images(self, info):
+    def resolve_images(self, *_args):
         return self.images.all()
 
     @classmethod
@@ -217,8 +218,9 @@ class Product(CountableDjangoObjectType):
         Image, description='The main thumbnail for a product.',
         size=graphene.Argument(graphene.Int, description='Size of thumbnail'))
     availability = graphene.Field(
-        ProductAvailability, description=dedent("""Informs about product's availability in the
-        storefront, current price and discounts."""))
+        ProductAvailability, description=dedent(
+            """Informs about product's availability in the
+               storefront, current price and discounts."""))
     price = graphene.Field(
         Money,
         description=dedent("""The product's base price (without any discounts
@@ -289,7 +291,7 @@ class Product(CountableDjangoObjectType):
         alt = image.alt if image else None
         return Image(alt=alt, url=url)
 
-    def resolve_url(self, info):
+    def resolve_url(self, *_args):
         return self.get_absolute_url()
 
     @gql_optimizer.resolver_hints(
@@ -303,17 +305,17 @@ class Product(CountableDjangoObjectType):
 
     @gql_optimizer.resolver_hints(
         prefetch_related='product_type__product_attributes__values')
-    def resolve_attributes(self, info):
+    def resolve_attributes(self, *_args):
         attributes_qs = self.product_type.product_attributes.all()
         return resolve_attribute_list(self.attributes, attributes_qs)
 
     @permission_required('product.manage_products')
-    def resolve_purchase_cost(self, info):
+    def resolve_purchase_cost(self, *_args):
         purchase_cost, _ = get_product_costs_data(self)
         return purchase_cost
 
     @permission_required('product.manage_products')
-    def resolve_margin(self, info):
+    def resolve_margin(self, *_args):
         _, margin = get_product_costs_data(self)
         return Margin(margin[0], margin[1])
 
@@ -325,16 +327,16 @@ class Product(CountableDjangoObjectType):
             raise GraphQLError('Product image not found.')
 
     @gql_optimizer.resolver_hints(model_field='images')
-    def resolve_images(self, info, **kwargs):
+    def resolve_images(self, *_args, **_kwargs):
         return self.images.all()
 
-    def resolve_variants(self, info, **kwargs):
+    def resolve_variants(self, *_args, **_kwargs):
         return self.variants.all()
 
-    def resolve_collections(self, info):
+    def resolve_collections(self, *_args):
         return self.collections.all()
 
-    def resolve_available_on(self, info):
+    def resolve_available_on(self, *_args):
         return self.publication_date
 
     @classmethod
@@ -354,12 +356,6 @@ class ProductType(CountableDjangoObjectType):
         PrefetchingConnectionField(
             Product, description='List of products of this type.'),
         prefetch_related=prefetch_products)
-    product_attributes = gql_optimizer.field(
-        PrefetchingConnectionField(Attribute),
-        model_field='product_attributes')
-    variant_attributes = gql_optimizer.field(
-        PrefetchingConnectionField(Attribute),
-        model_field='variant_attributes')
     tax_rate = TaxRateType(description='A type of tax rate.')
     variant_attributes = graphene.List(
         Attribute, description='Variant attributes of that product type.')
@@ -375,13 +371,15 @@ class ProductType(CountableDjangoObjectType):
             'has_variants', 'id', 'is_digital', 'is_shipping_required', 'name',
             'weight']
 
-    def resolve_product_attributes(self, info, **kwargs):
+    @gql_optimizer.resolver_hints(prefetch_related='product_attributes')
+    def resolve_product_attributes(self, *_args, **_kwargs):
         return self.product_attributes.all()
 
-    def resolve_variant_attributes(self, info, **kwargs):
+    @gql_optimizer.resolver_hints(prefetch_related='variant_attributes')
+    def resolve_variant_attributes(self, *_args, **_kwargs):
         return self.variant_attributes.all()
 
-    def resolve_products(self, info, **kwargs):
+    def resolve_products(self, info, **_kwargs):
         if hasattr(self, 'prefetched_products'):
             return self.prefetched_products
         qs = self.products.visible_to_user(info.context.user)
@@ -417,7 +415,7 @@ class Collection(CountableDjangoObjectType):
         interfaces = [relay.Node]
         model = models.Collection
 
-    def resolve_background_image(self, info, size=None, **kwargs):
+    def resolve_background_image(self, info, size=None, **_kwargs):
         if self.background_image:
             return Image.get_adjusted(
                 image=self.background_image,
@@ -427,13 +425,13 @@ class Collection(CountableDjangoObjectType):
                 info=info,
             )
 
-    def resolve_products(self, info, **kwargs):
+    def resolve_products(self, info, **_kwargs):
         if hasattr(self, 'prefetched_products'):
             return self.prefetched_products
         qs = self.products.visible_to_user(info.context.user)
         return gql_optimizer.query(qs, info)
 
-    def resolve_published_date(self, info):
+    def resolve_published_date(self, *_args):
         return self.publication_date
 
     @classmethod
@@ -483,11 +481,11 @@ class Category(CountableDjangoObjectType):
         interfaces = [relay.Node]
         model = models.Category
 
-    def resolve_ancestors(self, info, **kwargs):
+    def resolve_ancestors(self, info, **_kwargs):
         qs = self.get_ancestors()
         return gql_optimizer.query(qs, info)
 
-    def resolve_background_image(self, info, size=None, **kwargs):
+    def resolve_background_image(self, info, size=None, **_kwargs):
         if self.background_image:
             return Image.get_adjusted(
                 image=self.background_image,
@@ -497,14 +495,14 @@ class Category(CountableDjangoObjectType):
                 info=info,
             )
 
-    def resolve_children(self, info, **kwargs):
+    def resolve_children(self, info, **_kwargs):
         qs = self.children.all()
         return gql_optimizer.query(qs, info)
 
-    def resolve_url(self, info):
+    def resolve_url(self, _info):
         return self.get_absolute_url()
 
-    def resolve_products(self, info, **kwargs):
+    def resolve_products(self, info, **_kwargs):
         # If the category has no children, we use the prefetched data.
         children = self.children.all()
         if not children and hasattr(self, 'prefetched_products'):