+++ saleor/static/dashboard-next/collections/urls.ts
@@ -1,4 +1,7 @@
+import { stringify as stringifyQs } from "qs";
 import * as urlJoin from "url-join";
+
+import { Dialog, Pagination } from "../types";
 
 const collectionSectionUrl = "/collections/";
 
@@ -6,23 +9,10 @@ export const collectionListPath = collectionSectionUrl;
 export const collectionListUrl = collectionSectionUrl;
 
 export const collectionPath = (id: string) => urlJoin(collectionSectionUrl, id);
-export const collectionUrl = (id: string) =>
-  collectionPath(encodeURIComponent(id));
+export type CollectionUrlDialog = "remove" | "removeImage" | "assign";
+export type CollectionUrlQueryParams = Pagination & Dialog<CollectionUrlDialog>;
+export const collectionUrl = (id: string, params?: CollectionUrlQueryParams) =>
+  collectionPath(encodeURIComponent(id)) + "?" + stringifyQs(params);
 
 export const collectionAddPath = urlJoin(collectionSectionUrl, "add");
 export const collectionAddUrl = collectionAddPath;
-
-export const collectionRemovePath = (id: string) =>
-  urlJoin(collectionPath(id), "remove");
-export const collectionRemoveUrl = (id: string) =>
-  collectionRemovePath(encodeURIComponent(id));
-
-export const collectionImageRemovePath = (id: string) =>
-  urlJoin(collectionPath(id), "removeImage");
-export const collectionImageRemoveUrl = (id: string) =>
-  collectionImageRemovePath(encodeURIComponent(id));
-
-export const collectionAddProductPath = (id: string) =>
-  urlJoin(collectionPath(id), "add");
-export const collectionAddProductUrl = (id: string) =>
-  collectionAddProductPath(encodeURIComponent(id));